+++ backend-v3/libs/elasticsearch.js
@@ -9,8 +9,10 @@ const elasticsearchClient = new elasticsearch.Client(elasticsearchConfig)
 ready.addModule('elasticsearch')
 elasticsearchClient.ping({
   requestTimeout: 15000
-}).then(() => {
-  ready.ready('elasticsearch')
 })
+  .then(() => {
+    ready.ready('elasticsearch')
+  })
+  .catch((err) => console.error(err))
 
 module.exports = elasticsearchClient